Bumped version.
Made reel return an uint64_t for duration instead of size.
Fixed CMakeLists.txt so that SOURCES does not contain any .idl or .rc file in
it before xgettext gets called.

+++ mrViewer/src/gui/mrvReel.cpp
@@ -40,7 +40,7 @@ int64_t Reel_t::maximum() const
     return m->position() + m->duration() - 1;
 }
 
-size_t Reel_t::duration() const
+uint64_t Reel_t::duration() const
 {
     return maximum() - minimum() + 1;
 }
